Update UI and components for project funding comments.
Addresses https://geyserteam.atlassian.net/browse/GT-912
